feat: 대진표 생성 기능 구현

+++ badminton-api/src/main/java/org/badminton/api/league/model/dto/LeagueCreateRequest.java
@@ -39,7 +39,7 @@ public record LeagueCreateRequest(
 	LocalDateTime closedAt,
 
 	@Schema(description = "참가 인원", example = "16")
-	Long playerCount,
+	int playerCount,
 
 	@Schema(description = "매칭 조건", example = "TIER")
 	String matchingRequirement
