+++ badminton-api/src/main/java/org/badminton/api/league/model/dto/LeagueReadResponse.java
@@ -35,7 +35,7 @@ public record LeagueReadResponse(
 	LocalDateTime closedAt,
 
 	@Schema(description = "참가 인원", example = "16")
-	Long playerCount,
+	int playerCount,
 
 	@Schema(description = "생성 일자", example = "2024-09-10T15:30:00")
 	LocalDateTime createdAt,
